+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/listener/ProxyListenerRegister.java
@@ -17,7 +17,7 @@
 
 package io.shardingsphere.proxy.listener;
 
-import io.shardingsphere.transaction.common.listener.TransactionListener;
+import io.shardingsphere.transaction.listener.TransactionListener;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
