+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/TransactionTypeHolder.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.common;
+package io.shardingsphere.transaction;
 
 import io.shardingsphere.core.constant.TransactionType;
