+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/listener/TransactionListener.java
@@ -15,15 +15,15 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.common.listener;
+package io.shardingsphere.transaction.listener;
 
 import com.google.common.eventbus.AllowConcurrentEvents;
 import com.google.common.eventbus.Subscribe;
 import io.shardingsphere.core.util.EventBusInstance;
 import io.shardingsphere.transaction.manager.ShardingTransactionManager;
 import io.shardingsphere.transaction.manager.ShardingTransactionManagerRegistry;
-import io.shardingsphere.transaction.common.TransactionTypeHolder;
-import io.shardingsphere.transaction.common.event.TransactionEvent;
+import io.shardingsphere.transaction.TransactionTypeHolder;
+import io.shardingsphere.transaction.event.TransactionEvent;
 
 import java.sql.SQLException;
